% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export-pivot}
\alias{export-pivot}
\alias{export}
\alias{export_pptx}
\alias{export_docx}
\alias{export_xlsx}
\title{Export pivot table}
\usage{
export(x, output, to = c("pptx", "docx", "xlsx"), ...)

export_pptx(x, output, ...)

export_docx(x, output, ...)

export_xlsx(x, output, ...)
}
\arguments{
\item{x}{A \code{\link{pivot_table}} or \code{\link{pivot_format}} object.}

\item{output}{Path where to generate file.}

\item{to}{Output document: \code{"pptx"} (PowerPoint), \code{"docx"} (Word) or \code{"xlsx"} (Excel).}

\item{...}{Arguments passed to \code{\link{pivot_format}} if \code{x} isn't formated yet.}
}
\value{
Path to output (invisibly).
}
\description{
Export formated pivot table to PowerPoint, Word or Excel document.
}
\examples{
library(flexpivot)
data("nobel_laureates")

pt <- pivot_table(nobel_laureates, rows = "category", cols = "gender")

# To PowerPoint
path_pptx <- tempfile(fileext = ".pptx")
export_pptx(pt, path_pptx)
browseURL(path_pptx)


# To Word
path_docx <- tempfile(fileext = ".docx")
export_docx(pt, path_docx)
browseURL(path_docx)


# To Excel
path_xlsx <- tempfile(fileext = ".xlsx")
export_xlsx(pt, path_xlsx)
browseURL(path_xlsx)


# Clean up
unlink(path_pptx)
unlink(path_docx)
unlink(path_xlsx)
}
