% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex}
\alias{flex}
\title{Flex format for \code{pivot_table}}
\usage{
flex(pivot, background = "#81A1C1")
}
\arguments{
\item{pivot}{a \code{pivot_table} object.}

\item{background}{Background color.}
}
\value{
a \code{flextable} object.
}
\description{
Flex format for \code{pivot_table}
}
\examples{

data("diamonds", package = "ggplot2")

# With two variables
flex(pivot_table(diamonds, rows = "cut", cols = "color"))

# Only count
flex(pivot_table(diamonds, rows = "cut", cols = "color", stats = "n"))

# Only percent
flex(pivot_table(diamonds, rows = "cut", cols = "color", stats = "p"))



# Two variable as rows
flex(pivot_table(diamonds, rows = c("clarity", "cut"), cols = "color"))

# Two variable as cols
flex(pivot_table(diamonds, rows = "cut", cols = c("color", "clarity")))


# Without cols
flex(pivot_table(diamonds, rows = "cut"))
flex(pivot_table(diamonds, rows = c("clarity", "cut")))
}
