% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{pivot-shiny}
\alias{pivot-shiny}
\alias{pivot2Output}
\alias{renderPivot2}
\alias{pivotOutput}
\alias{export_labels}
\alias{renderPivot}
\title{Pivot table Output in Shiny}
\usage{
pivot2Output(outputId, width = "100\%", ...)

renderPivot2(
  expr,
  width = 1,
  background = "#81A1C1",
  color = "#FFFFFF",
  border = "#FFFFFF",
  font_size = 14,
  font_name = NULL,
  labels = pivot_labels(),
  formatter = pivot_formatter(),
  env = parent.frame(),
  quoted = FALSE
)

pivotOutput(outputId, width = "100\%", export = export_labels(), ...)

export_labels(
  export = "Export",
  clipboard = "Copy to clipboard",
  powerpoint = "Export to PowerPoint",
  word = "Export to Word",
  excel = "Export to Excel"
)

renderPivot(
  expr,
  width = 1,
  background = "#81A1C1",
  color = "#FFFFFF",
  border = "#FFFFFF",
  font_size = 14,
  font_name = NULL,
  labels = pivot_labels(),
  formatter = pivot_formatter(),
  label_value = "value",
  env = parent.frame(),
  quoted = FALSE,
  filename = "export-pivot"
)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width}{Value of the preferred width of the table in percent (\code{[0,1]}).}

\item{...}{Other arguments to pass to the container tag function.
This is useful for providing additional classes for the tag.}

\item{expr}{An expression that generates a \code{\link{pivot_table}}.}

\item{background}{Background color for the header and column(s) containing row(s).}

\item{color}{Text color for the header and column(s) containing row(s).}

\item{border}{Border color (applies to all table).}

\item{font_size}{Font size (applies to all table).}

\item{font_name}{Font name (applies to all table).}

\item{labels}{Custom labels for statistics, see \code{\link{pivot_labels}}.}

\item{formatter}{Function to format content, see \code{\link{pivot_formatter}}.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{export}{Export labels, use \code{NULL} to disable all exports.}

\item{clipboard, powerpoint, word, excel}{Labels to display in
export menu, use \code{NULL} to disable specific format.}

\item{label_value}{For Excel output, the label for variable containing the values.}

\item{filename}{A string of the filename to export WITHOUT extension, it will be added accordint to type of export.}
}
\value{
An HTML output element that can be included in Shiny UI.
}
\description{
Display a \code{\link{pivot_table}} in Shiny:
\itemize{
\item pivotOutput / renderPivot: adds a button to download pivot table in PowerPoint, Word and Excel.
\item pivot2Output / renderPivot2: display only the pivot table.
}
}
\examples{
library(shiny)
library(flexpivot)
data("nobel_laureates")

ui <- fluidPage(
  tags$h2("Pivot table in Shiny"),
  fluidRow(
    column(
      width = 6,
      selectInput(
        inputId = "row",
        label = "Row",
        choices = c("category", "gender", "birth_continent", "laureate_type"),
        width = "100\%"
      )
    ),
    column(
      width = 6,
      selectInput(
        inputId = "col",
        label = "Col",
        choices = c("category", "gender", "birth_continent", "laureate_type"),
        selected = "gender",
        width = "100\%"
      )
    )
  ),
  pivotOutput("pivot")
)

server <- function(input, output, session) {

  output$pivot <- renderPivot({
    pivot_table(nobel_laureates, input$row, input$col)
  }, background = "#A3BE8C")

}

if (interactive())
  shinyApp(ui, server)
}
