% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_addin.R
\name{pivot_addin}
\alias{pivot_addin}
\title{Pivot table RStudio addin}
\usage{
pivot_addin(data)
}
\arguments{
\item{data}{a \code{data.frame}.}
}
\value{
No value.
}
\description{
Pivot table RStudio addin
}
\examples{
if (interactive()) {

  library(flexpivot)
  data("nobel_laureates")
  pivot_addin(nobel_laureates)

}
}
