% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_format.R
\name{pivot_format}
\alias{pivot_format}
\title{Flextable format for \code{pivot_table}}
\usage{
pivot_format(
  pivot,
  background = "#81A1C1",
  color = "#FFFFFF",
  border = "#FFFFFF",
  fontSize = 14,
  labels = pivot_labels(),
  formatter = pivot_formatter(),
  drop_stats = FALSE,
  keep_data = TRUE
)
}
\arguments{
\item{pivot}{A \code{\link{pivot_table}} object.}

\item{background}{Background color for the header and column(s) containing row(s).}

\item{color}{Text color for the header and column(s) containing row(s).}

\item{border}{Border color (applies to all table).}

\item{fontSize}{Font size (applies to all table).}

\item{labels}{Custom labels for statistics, see \code{\link{pivot_labels}}.}

\item{formatter}{Function to format content, see \code{\link{pivot_formatter}}.}

\item{drop_stats}{Drop the stats column, can be useful if have only one stat to show.}

\item{keep_data}{Keep data as attribute, this can
be useful to retrieve the data from which the table was formatted.}
}
\value{
a \code{flextable} object.
}
\description{
Flextable format for \code{pivot_table}
}
\examples{
library(flexpivot)
library(magrittr)
data("nobel_laureates")

# With two variables
nobel_laureates \%>\%
  pivot_table("category", "gender") \%>\%
  pivot_format()

# Only count
nobel_laureates \%>\%
  pivot_table("category", "gender", stats = "n") \%>\%
  pivot_format(drop_stats = TRUE)

# Only percent
nobel_laureates \%>\%
  pivot_table("category", "gender", stats = "p") \%>\%
  pivot_format()


# Two variable as rows
nobel_laureates \%>\%
  pivot_table(c("birth_continent", "gender"), "category", stats = "p") \%>\%
  pivot_format()


# Without cols
pivot_format(pivot_table(nobel_laureates, rows = "category"))
pivot_format(pivot_table(nobel_laureates, rows = c("category", "gender")))


# Customize
nobel_laureates \%>\%
  pivot_table("category") \%>\%
  pivot_format(
    background = "#A3BE8C",
    labels = pivot_labels(n = "Count", p = "Percentage")
  )







}
