% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{Create a pivot table}
\usage{
pivot_table(
  data,
  rows,
  cols = NULL,
  wt = NULL,
  stats = c("n", "p", "p_row", "p_col"),
  digits = 2,
  total = TRUE,
  total_label = "Total",
  na_label = "<missing>",
  complete = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame}.}

\item{rows}{Character vector of variable(s) to use as rows.}

\item{cols}{Character vector of variable(s) to use as columns.}

\item{wt}{Character, variable to use as weights if any.}

\item{stats}{Statistic(s) to compute.}

\item{digits}{Integer indicating the number of decimal places to be used.}

\item{total}{Logical, add total or not.}

\item{total_label}{Label to use fo total.}

\item{na_label}{Label to use for missing values.}

\item{complete}{Complete missing combination between
\code{rows} if several and \code{cols} if any.}
}
\value{
a \code{data.table}
}
\description{
Create a pivot table
}
\examples{
library(flexpivot)
data("nobel_laureates")

# One variable
pivot_table(nobel_laureates, rows = "category")

# With two variables
pivot_table(nobel_laureates, rows = "category", cols = "birth_continent")

# Only count
pivot_table(nobel_laureates, rows = "category", cols = "gender", stats = "n")

# Only percent
pivot_table(nobel_laureates, rows = "category", cols = "gender", stats = "p")

# Without total
pivot_table(nobel_laureates, rows = "category", cols = "gender", total = FALSE)


# Two variable as rows
pivot_table(nobel_laureates, rows = c("birth_continent", "category"), cols = "gender")

# Two variable as cols
pivot_table(nobel_laureates, rows = "category", cols = c("gender", "laureate_type"))


# Round
pivot_table(nobel_laureates, rows = "category", digits = 0)

# Without cols
pivot_table(nobel_laureates, rows = "category")
pivot_table(nobel_laureates, rows = c("category", "birth_continent"))
}
