% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot}
\alias{unpivot}
\title{Unpivot}
\usage{
unpivot(x)
}
\arguments{
\item{x}{A \code{\link{pivot_table}} object.}
}
\value{
a \code{data.table}.
}
\description{
From a \code{\link{pivot_table}} object put column(s) back as rows.
}
\examples{
library(flexpivot)
library(magrittr)
data("nobel_laureates")

# Revert format
nobel_laureates \%>\%
  pivot_table(
    rows = "category",
    cols = c("gender", "birth_continent")
  ) \%>\%
  unpivot()

nobel_laureates \%>\%
  pivot_table(
    rows = "category",
    cols = "birth_continent"
  ) \%>\%
  unpivot()

}
