% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export-pivot}
\alias{export-pivot}
\alias{export}
\alias{export_pptx}
\alias{export_docx}
\title{Export pivot table}
\usage{
export(x, output, to = c("pptx", "docx", "xlsx"), ...)

export_pptx(x, output, ...)

export_docx(x, output, ...)
}
\arguments{
\item{x}{A \code{\link{pivot_table}} or \code{\link{pivot_format}} object.}

\item{output}{Path where to generate file.}

\item{to}{Output document: \code{"pptx"} (PowerPoint), \code{"docx"} (Word) or \code{"xlsx"} (Excel).}

\item{...}{Arguments passed to \code{\link{pivot_format}} if \code{x} isn't formated yet.}
}
\value{
Path to output (invisibly).
}
\description{
Export formated pivot table to PowerPoint, Word or Excel document.
}
\examples{
library(flexpivot)
data("diamonds", package = "ggplot2")

pt <- pivot_table(diamonds, rows = "cut", cols = "color")

# To PowerPoint
path_pptx <- tempfile(fileext = ".pptx")
export_pptx(pt, path_pptx)
browseURL(path_pptx)


# To Word
path_docx <- tempfile(fileext = ".docx")
export_docx(pt, path_docx)
browseURL(path_docx)



# Clean up
unlink(path_pptx)
unlink(path_docx)
}
