% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_format.R
\name{pivot_format}
\alias{pivot_format}
\title{Flextable format for \code{pivot_table}}
\usage{
pivot_format(
  pivot,
  background = "#81A1C1",
  border = "#FFFFFF",
  fontSize = 14,
  labels = pivot_labels(),
  formatter = pivot_formatter()
)
}
\arguments{
\item{pivot}{a \code{pivot_table} object.}

\item{background}{Background color.}

\item{border}{Border color (applies to all table).}

\item{fontSize}{Font size (applies to all table).}

\item{labels}{Custom labels for statistics, see \code{\link{pivot_labels}}.}

\item{formatter}{Function to format content, see \code{\link{pivot_formatter}}.}
}
\value{
a \code{flextable} object.
}
\description{
Flextable format for \code{pivot_table}
}
\examples{

data("diamonds", package = "ggplot2")

# With two variables
pivot_format(pivot_table(diamonds, rows = "cut", cols = "color"))

# Only count
pivot_format(pivot_table(diamonds, rows = "cut", cols = "color", stats = "n"))

# Only percent
pivot_format(pivot_table(diamonds, rows = "cut", cols = "color", stats = "p"))



# Two variable as rows
pivot_format(pivot_table(diamonds, rows = c("clarity", "cut"), cols = "color"))

# Two variable as cols
pivot_format(pivot_table(diamonds, rows = "cut", cols = c("color", "clarity")))


# Without cols
pivot_format(pivot_table(diamonds, rows = "cut"))
pivot_format(pivot_table(diamonds, rows = c("clarity", "cut")))
}
