% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_format.R
\name{pivot_labels}
\alias{pivot_labels}
\title{Labels for \code{pivot_format}}
\usage{
pivot_labels(
  stats = "Statistic",
  n = "N",
  p = "\%",
  p_col = "Col \%",
  p_row = "Row \%",
  rows = NULL
)
}
\arguments{
\item{stats}{Name of statistics column.}

\item{n}{Count.}

\item{p}{Percentage.}

\item{p_col}{Column perc.}

\item{p_row}{Row perc.}

\item{rows}{Labels for variables use as rows.}
}
\value{
a \code{list} that can be use in \code{\link{pivot_format}}.
}
\description{
Labels for \code{pivot_format}
}
\examples{
data("diamonds", package = "ggplot2")

# With two variables
pivot_format(
  pivot = pivot_table(diamonds, rows = "cut", cols = "color"),
  labels = pivot_labels(
    stats = "Statistique",
    n = "N",
    p = "\%",
    rows = c("Cut variable")
  )
)
}
