% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{Create a pivot table}
\usage{
pivot_table(
  data,
  rows,
  cols = NULL,
  wt = NULL,
  stats = c("n", "p", "p_row", "p_col"),
  total = TRUE,
  total_label = "Total",
  complete = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame}.}

\item{rows}{Character vector of variable(s) to use as rows.}

\item{cols}{Character vector of variable(s) to use as columns.}

\item{wt}{Character, variable to use as weights if any.}

\item{stats}{Statistic(s) to compute.}

\item{total}{Logical, add total or not.}

\item{total_label}{Label to use fo total.}

\item{complete}{Complete missing combination between
\code{rows} if several and \code{cols} if any.}
}
\value{
a \code{data.table}
}
\description{
Create a pivot table
}
\examples{

data("diamonds", package = "ggplot2")

# With two variables
pivot_table(diamonds, rows = "cut", cols = "color")

# Only count
pivot_table(diamonds, rows = "cut", cols = "color", stats = "n")

# Only percent
pivot_table(diamonds, rows = "cut", cols = "color", stats = "p")

# Without total
pivot_table(diamonds, rows = "cut", cols = "color", total = FALSE)


# Two variable as rows
pivot_table(diamonds, rows = c("clarity", "cut"), cols = "color")

# Two variable as cols
pivot_table(diamonds, rows = "cut", cols = c("color", "clarity"))


# Without cols
pivot_table(diamonds, rows = "cut")
pivot_table(diamonds, rows = c("clarity", "cut"))
}
