% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpivot.R
\name{unpivot}
\alias{unpivot}
\title{Unpivot}
\usage{
unpivot(x)
}
\arguments{
\item{x}{A \code{\link{pivot_table}} object.}
}
\value{
a \code{data.table}.
}
\description{
From a \code{\link{pivot_table}} object put column(s) back as rows.
}
\examples{
library(flexpivot)
data("diamonds", package = "ggplot2")

# Revert format
unpivot(pivot_table(diamonds, rows = "cut", cols = c("color", "clarity")))
unpivot(pivot_table(diamonds, rows = "cut", cols = c("color")))
}
